refactor: rename filenames and other names
