+++ setup_wizard/tests/config.json.sample
@@ -1,6 +1,6 @@
 {
     "blender_execution_file_path": "FILE_PATH_TO_blender.exe",
     "festivity_root_folder_file_path": "FILE_PATH_TO_festivity_shaders_folder",
-    "characters_folder_file_path": "FILE_PATH_TO_gi-assets-characters-folder",
+    "characters_folder_file_path": "FILE_PATH_TO_characters-folder",
     "material_json_folder_file_path": "FILE_PATH_TO_material_data_folder"
 }
